Combined the bridge scripts into one to prepare it for a stateless deployment.
